requirements.txt
matplotlib==3.3.3
uvicorn==0.13.2
starlette==0.13.6
tensorflow==2.4.0
numpy==1.19.4
fastapi==0.63.0
Pillow==8.0.1
